[FIRRTL] Move MemToRegOfVec implementation to FullReset
