+++ lib/Dialect/FIRRTL/Transforms/CMakeLists.txt
@@ -46,7 +46,6 @@ add_circt_dialect_library(CIRCTFIRRTLTransforms
   LowerTypes.cpp
   LowerXMR.cpp
   MaterializeDebugInfo.cpp
-  MemToRegOfVec.cpp
   MergeConnections.cpp
   ModuleInliner.cpp
   ModuleSummary.cpp
